Improve Moments creation
