+++ packages/moments/src/client/dtos/create/CreateSteps.ts
@@ -1,0 +1,9 @@
+export enum CreateSteps {
+  REQUESTING_MEDIA_UPLOAD_URL,
+  UPLOADING_MEDIA,
+  UPLOADING_MEDIA_METADATA,
+  PROCESSING_MEDIA,
+  PROCESSING_MEDIA_ERROR,
+  UPLOADING_MOMENT,
+  FINISHED,
+}
