+++ packages/moments/src/client/dtos/fetch/FetchMomentsInput.ts
@@ -1,23 +1,4 @@
 /**
- * Interface representing the input needed to create a moment.
- * @interface
- * @property {number} dropId - The ID of the drop related to the moment.
- * @property {number} [tokenId] - The ID of the token related to the moment (optional).
- * @property {Buffer} file - The buffer containing the file data.
- * @property {string} author - The author of the moment.
- * @property {string} timeOut - The amount of time to wait until media is processed.
- * @property {string} fileType - The type of the file.
- */
-export interface CreateMomentInput {
-  dropId: number;
-  tokenId?: number;
-  file: Buffer;
-  author: string;
-  timeOut?: number;
-  fileType: string;
-}
-
-/**
  * Interface representing the input needed to fetch moments.
  * @interface
  * @property {number} limit - The maximum number of moments to fetch.
