+++ packages/moments/src/index.ts
@@ -1,3 +1,5 @@
-export { CreateMomentInput, FetchMomentsInput } from './types/input';
 export { Moment } from './domain/Moment';
-export { MomentsClient } from './MomentsClient';
+export { MomentsClient } from './client/MomentsClient';
+export { CreateMomentInput } from './client/dtos/create/CreateInput';
+export { CreateSteps } from './client/dtos/create/CreateSteps';
+export { FetchMomentsInput } from './client/dtos/fetch/FetchMomentsInput';
